API

BaseModelica.BaseModelicaModel — Type

Represents a BaseModelica model.

BaseModelica.BaseModelicaPackage — Type

Holds the name of the package, the models in the package, and eventually BaseModelica records.

BaseModelica.parse_file — Method

Takes a path to a file and parses the contents in to a BaseModelicaPackage.

BaseModelica.parse_str — Method

Parses a string in to a BaseModelicaPackage.